refactor tests to new class
